solution for issue #5997
I add this file related to issue #5981 about GC.
And for the issue #5997 , I add lines to update and get session id from what saved in DB (as I discussed on a comment in the issue).

I fased losing session when mobile app send request which seem to redirect. And this solve lossing or each updating session.

Description :
The main concept of my solution is to get session id when user logged successfully and save it in the database. Then ask to return session id from DB each time when system checks user authenticate. After that, use it to update codeigniter session for user by set it as current session and terminate it when user logged out or not using the account for a period of time.